+++ Controllers/UsuariosController.php
@@ -24,14 +24,18 @@ class UsuariosController extends AbstractController {
      * Recupera todos os registros
      */
     protected function findAll() {
-        $viewModel = new UsuariosModel();
         $url = $this->request["url"];
-        $res = $viewModel->findAll();
 
-        if (!$res) {
-            $res = array();
-        } else {
-            $this->returnJson($res, OK_CODE);
+        //valida quem tem acesso a esse metodo
+        if ($this->authorization([ADMINISTRADOR], $this->request["authorization"], $url)) {
+            $viewModel = new UsuariosModel();
+            $res = $viewModel->findAll();
+
+            if (!$res) {
+                $res = array();
+            } else {
+                $this->returnJson($res, OK_CODE);
+            }
         }
     }
 
@@ -39,43 +43,54 @@ class UsuariosController extends AbstractController {
      * Recupera um registro
      */
     protected function findById() {
-        $viewModel = new UsuariosModel();
         $url = $this->request["url"];
-        $res = $viewModel->findById($this->request['id']);
 
-        $code = OK_CODE;
-        if (!$res) {
-            $res = array();
-            $code = NOT_FOUND_CODE;
+        //valida quem tem acesso a esse metodo
+        if ($this->authorization([ADMINISTRADOR], $this->request["authorization"], $url)) {
+            $viewModel = new UsuariosModel();
+            $res = $viewModel->findById($this->request['id']);
+
+            $code = OK_CODE;
+            if (!$res) {
+                $res = array();
+                $code = NOT_FOUND_CODE;
+            }
+            $this->returnJson($res, $code);
         }
-        $this->returnJson($res, $code);
     }
 
     /**
      * ADD share
      */
     protected function login() {
-        $viewModel = new UsuariosModel();
         $url = $this->request["url"];
-        $body = (array) $this->request["body"];
-        $obj = new Usuarios();
 
-        if (isset($body['email']) && isset($body['senha'])) {
-            $obj->setEmail($body['email']);
-            $obj->setSenha($body['senha']);
+        //verifica se está sendo acessado pelo metodo POST
+        if ($this->request['type'] == "POST") {
 
-            $token = $viewModel->login($obj);
+            $viewModel = new UsuariosModel();
+            $body = (array) $this->request["body"];
+            $obj = new Usuarios();
 
-            if ($token == ERROR_LOGIN_EMAIL) {
-                $res = new StandartError(NOT_FOUND_CODE, NOT_FOUND, ERROR_LOGIN_EMAIL, $url);
-                $res->getJsonError();
-            } else if ($token == ERROR_LOGIN_SENHA) {
-                $res = new StandartError(NOT_FOUND_CODE, NOT_FOUND, ERROR_LOGIN_SENHA, $url);
-                $res->getJsonError();
-            } else {
+            if (isset($body['email']) && isset($body['senha'])) {
+                $obj->setEmail($body['email']);
+                $obj->setSenha($body['senha']);
 
-                $this->returnJsonLogin($token, OK_CODE);
+                $token = $viewModel->login($obj);
+
+                if ($token == ERROR_LOGIN_EMAIL) {
+                    $res = new StandartError(NOT_FOUND_CODE, NOT_FOUND, ERROR_LOGIN_EMAIL, $url);
+                    $res->getJsonError();
+                } else if ($token == ERROR_LOGIN_SENHA) {
+                    $res = new StandartError(NOT_FOUND_CODE, NOT_FOUND, ERROR_LOGIN_SENHA, $url);
+                    $res->getJsonError();
+                } else {
+                    $this->returnJsonLogin($token, OK_CODE);
+                }
             }
+        } else {
+            $res = new StandartError(METHOD_NOT_ALLOWED_CODE, METHOD_NOT_ALLOWED, METHOD_NOT_ALLOWED_MSG, $url);
+            $res->getJsonError();
         }
     }
 
@@ -83,20 +98,24 @@ class UsuariosController extends AbstractController {
      * ADD share
      */
     protected function add() {
-        $viewModel = new UsuariosModel();
         $url = $this->request["url"];
-        $body = (array) $this->request["body"];
 
-        $obj = new Usuarios();
+        //valida quem tem acesso a esse metodo
+        if ($this->authorization([ADMINISTRADOR], $this->request["authorization"], $url)) {
+            $viewModel = new UsuariosModel();
+            $body = (array) $this->request["body"];
 
-        if ($obj->validaCampos($obj->getCampos(), $body, $url)) {
+            $obj = new Usuarios();
 
-            $obj->setId(null);
-            $obj->setEmail($body['email']);
-            $obj->setSenha($body['senha']);
-            $obj->setAcesso(new Acessos());
-            $obj->getAcesso()->setId($body['acessos_id']);
-            $this->returnJson($viewModel->add($obj), CREATE_CODE, $url);
+            if ($obj->validaCampos($obj->getCampos(), $body, $url)) {
+
+                $obj->setId(null);
+                $obj->setEmail($body['email']);
+                $obj->setSenha($body['senha']);
+                $obj->setAcesso(new Acessos());
+                $obj->getAcesso()->setId($body['acessos_id']);
+                $this->returnJson($viewModel->add($obj), CREATE_CODE, $url);
+            }
         }
     }
 
@@ -104,31 +123,35 @@ class UsuariosController extends AbstractController {
      * Atualiza um registro
      */
     protected function update() {
-        $viewModel = new UsuariosModel();
         $url = $this->request["url"];
 
-        $obj = new Usuarios();
+        //valida quem tem acesso a esse metodo
+        if ($this->authorization([ADMINISTRADOR], $this->request["authorization"], $url)) {
+            $viewModel = new UsuariosModel();
 
-        $id = $this->request["id"];
-        $obj_old = $viewModel->findById($id);
-        if ($obj_old) {
+            $obj = new Usuarios();
 
-            $obj_new = (array) $this->request["body"];
-            $body = $obj->compareDif($obj_new, $obj_old, $obj->getCampos());
+            $id = $this->request["id"];
+            $obj_old = $viewModel->findById($id);
+            if ($obj_old) {
 
-            $code = OK_CODE;
+                $obj_new = (array) $this->request["body"];
+                $body = $obj->compareDif($obj_new, $obj_old, $obj->getCampos());
 
-            if ($obj->validaCampos($obj->getCampos(), $body, $url, true)) {
-                $obj->setId($body['id']);
-                $obj->setEmail($body['email']);
-                $obj->setAcesso(new Acessos());
-                $obj->getAcesso()->setId($body['acessos_id']);
+                $code = OK_CODE;
 
-                $this->returnJson($viewModel->update($obj), $code);
+                if ($obj->validaCampos($obj->getCampos(), $body, $url, true)) {
+                    $obj->setId($body['id']);
+                    $obj->setEmail($body['email']);
+                    $obj->setAcesso(new Acessos());
+                    $obj->getAcesso()->setId($body['acessos_id']);
+
+                    $this->returnJson($viewModel->update($obj), $code);
+                }
+            } else {
+                $res = new StandartError(BAD_REQUEST_CODE, NOT_FOUND, NOT_FOUND_ID, $url);
+                $res->getJsonError();
             }
-        } else {
-            $res = new StandartError(BAD_REQUEST_CODE, NOT_FOUND, NOT_FOUND_ID, $url);
-            $res->getJsonError();
         }
     }
 
@@ -137,15 +160,19 @@ class UsuariosController extends AbstractController {
      */
     protected function delete() {
         $url = $this->request["url"];
-        $viewModel = new UsuariosModel();
-        $obj = $viewModel->findById($this->request['id']);
 
-        if ($obj) {
-            $res = $viewModel->delete($this->request['id']);
-            $this->returnJson($res, OK_CODE);
-        } else {
-            $res = new StandartError(BAD_REQUEST_CODE, NOT_FOUND, NOT_FOUND_ID, $url);
-            $res->getJsonError();
+        //valida quem tem acesso a esse metodo
+        if ($this->authorization([ADMINISTRADOR], $this->request["authorization"], $url)) {
+            $viewModel = new UsuariosModel();
+            $obj = $viewModel->findById($this->request['id']);
+
+            if ($obj) {
+                $res = $viewModel->delete($this->request['id']);
+                $this->returnJson($res, OK_CODE);
+            } else {
+                $res = new StandartError(BAD_REQUEST_CODE, NOT_FOUND, NOT_FOUND_ID, $url);
+                $res->getJsonError();
+            }
         }
     }
